Bold new resource group in all labs
Students continuously miss that the need for a new resource group in these labs based on the way it is written. I also missed this during my first few walkthroughs. If we could simple bold new resource group or make the users more aware in the steps that a new resource group is required