Response to Issue #559 in original repo
Ex 1 Task 2 Step 5 Note 3
The steps in the note work as written, but it is not clear what action needs to be taken if a learner has chosen a location that they cannot deploy into.  I added a clarification that explains what will need to be done if they have chosen a location that cannot have the appropriate VM sku deployed.